Fix signed comparison on gbz80, fix comparison in pointer location for memmove().
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@7232 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/device/lib/_memmove.c
@@ -28,6 +28,7 @@
    might be covered by the GNU General Public License.
 -------------------------------------------------------------------------*/
 #include "string.h" 
+#include <stdint.h>
 #include <sdcc-lib.h>
 
 #ifndef _SDCC_PORT_PROVIDES_MEMMOVE
@@ -46,7 +47,7 @@ void * memmove (
 	char * d;
 	char * s;
 
-	if ((int)src < (int)dst) {
+	if ((uintptr_t)src < (uintptr_t)dst) {
 		/*
 		 * copy from higher addresses to lower addresses
 		 */
